Weaken the "suck in leftmost dummies" pass
Sometimes dummies are already to the right of their destination node.
That's fine, it doesn't actually hurt layout clarity. But the "suck in"
pass would actually force dummies very far to the left to avoid this
"problem" and would end up pushing other dummies off the left side of
the screen. Dumb!